Move getTeamOrThrow func utils -> team repo
